modification of VehicleOccupancies Oditi to match the new structure
