+++ internal/vehicleoccupancies/contextGtfsRt.go
@@ -103,12 +103,17 @@ func (d *VehicleOccupanciesGtfsRtContext) AddVehicleOccupancy(vehicleoccupancy *
 }
 
 /********* INTERFACE METHODS IMPLEMENTS *********/
+func (d *VehicleOccupanciesGtfsRtContext) InitContext(filesURI, externalURI url.URL,
+	externalToken string, navitiaURI url.URL, navitiaToken string, loadExternalRefresh,
+	connectionTimeout time.Duration, location *time.Location, occupancyActive bool) {
+
+	d.voContext = &VehicleOccupanciesContext{}
+	d.voContext.ManageVehicleOccupancyStatus(occupancyActive)
+}
 
 func (d *VehicleOccupanciesGtfsRtContext) RefreshVehicleOccupanciesLoop(externalURI url.URL,
 	externalToken string, navitiaURI url.URL, navitiaToken string, loadExternalRefresh,
 	connectionTimeout time.Duration, location *time.Location) {
-
-	d.voContext = &VehicleOccupanciesContext{}
 
 	refreshVehicleOccupanciesLoop(d, externalURI, externalToken, navitiaURI, navitiaToken,
 		loadExternalRefresh, connectionTimeout)
