+++ internal/vehicleoccupancies/navitia.go
@@ -10,6 +10,7 @@ import (
 	"net/url"
 	"time"
 
+	"github.com/CanalTP/forseti/internal/data"
 	"github.com/CanalTP/forseti/internal/utils"
 )
 
@@ -17,6 +18,7 @@ const (
 	URL_GET_LAST_LOAD       = "%s/status?"
 	URL_GET_VEHICLE_JOURNEY = "%s/vehicle_journeys?filter=vehicle_journey.has_code(%s)&"
 	STOP_POINT_CODE         = "gtfs_stop_code" // type code vehicle journey Navitia, the same of stop_id from Gtfs-rt
+	URL_GET_ROUTES          = "%s/lines/%s/route_schedules?direction_type=%s&from_datetime=%s"
 )
 
 // Structure to load the last date of modification static data
@@ -123,6 +125,58 @@ func GetVehicleJourney(id_gtfsRt string, uri url.URL, token string, connectionTi
 	return CreateVehicleJourney(navitiaVJ, id_gtfsRt, time.Now()), nil
 }
 
+// Get all Routes for Oditi service, just lines 40 and 45
+func LoadRoutesWithDirection(startIndex int, uri url.URL, token, direction string,
+	connectionTimeout time.Duration, location *time.Location) ([]RouteSchedule, error) {
+	// Call and load routes for line=40
+	dateTime := time.Now().Truncate(24 * time.Hour).Format("20060102T000000")
+	callUrl := fmt.Sprintf(URL_GET_ROUTES, uri.String(), "line:0:004004040:40", direction, dateTime)
+	resp, err := CallNavitia(callUrl, token, connectionTimeout)
+	if err != nil {
+		VehicleOccupanciesLoadingErrors.Inc()
+		return nil, err
+	}
+
+	navitiaRoutes := &data.NavitiaRoutes{}
+	decoder := json.NewDecoder(resp.Body)
+	err = decoder.Decode(navitiaRoutes)
+	if err != nil {
+		VehicleOccupanciesLoadingErrors.Inc()
+		return nil, err
+	}
+
+	sens := 0
+	if direction == "backward" {
+		sens = 1
+	}
+	routeSchedules := LoadRouteSchedulesData(startIndex, navitiaRoutes, sens, location)
+
+	// Call and load routes for line=45
+	callUrl = fmt.Sprintf(URL_GET_ROUTES, uri.String(), "line:0:004004029:45", direction, dateTime)
+	resp, err = CallNavitia(callUrl, token, connectionTimeout)
+	if err != nil {
+		VehicleOccupanciesLoadingErrors.Inc()
+		return nil, err
+	}
+
+	navitiaRoutes = &data.NavitiaRoutes{}
+	decoder = json.NewDecoder(resp.Body)
+	err = decoder.Decode(navitiaRoutes)
+	if err != nil {
+		VehicleOccupanciesLoadingErrors.Inc()
+		return nil, err
+	}
+
+	startIndex += len(routeSchedules)
+
+	rs45 := LoadRouteSchedulesData(startIndex+1, navitiaRoutes, sens, location)
+	// Concat two arrays
+	for i := range rs45 {
+		routeSchedules = append(routeSchedules, rs45[i])
+	}
+	return routeSchedules, nil
+}
+
 // This method call Navitia api with specific url and return a request response
 func CallNavitia(callUrl string, token string, connectionTimeout time.Duration) (*http.Response, error) {
 	resp, err := utils.GetHttpClient(callUrl, token, "Authorization", connectionTimeout)
@@ -155,4 +209,31 @@ func CreateVehicleJourney(navitiaVJ *NavitiaVehicleJourney, id_gtfsRt string, cr
 	}
 	vj := NewVehicleJourney(navitiaVJ.VehicleJourneys[0].ID, id_gtfsRt, sp, createDate)
 	return vj
+}
+
+func LoadRouteSchedulesData(startIndex int, navitiaRoutes *data.NavitiaRoutes, direction int,
+	location *time.Location) []RouteSchedule {
+	// Read RouteSchedule response body in json
+	// Normally there is one vehiclejourney for each departure.
+	routeScheduleList := make([]RouteSchedule, 0)
+	lineCode := navitiaRoutes.RouteSchedules[0].DisplayInformations.Code
+	for i := 0; i < len(navitiaRoutes.RouteSchedules[0].Table.Rows); i++ {
+		for j := 0; j < len(navitiaRoutes.RouteSchedules[0].Table.Rows[i].DateTimes); j++ {
+			rs, err := NewRouteSchedule(
+				lineCode,
+				navitiaRoutes.RouteSchedules[0].Table.Rows[i].StopPoint.ID,
+				navitiaRoutes.RouteSchedules[0].Table.Rows[i].DateTimes[j].Links[0].Value,
+				navitiaRoutes.RouteSchedules[0].Table.Rows[i].DateTimes[j].DateTime,
+				direction,
+				startIndex,
+				(i == 0), // StopPoint of departure for a distinct vehiclejourney
+				location)
+			if err != nil {
+				continue
+			}
+			routeScheduleList = append(routeScheduleList, *rs)
+			startIndex++
+		}
+	}
+	return routeScheduleList
 }
